fix(php85):  non-canonical cast

+++ system/src/Settings.php
@@ -497,7 +497,7 @@ class Settings implements \ArrayAccess
 
 				case 'double':
 				case 'float':
-					$ret['value'] = (double)($ret['value']);
+					$ret['value'] = (float)($ret['value']);
 					break;
 
 				default:
